fix clippy

+++ optd/core/src/ir/operator/logical/get.rs
@@ -47,7 +47,7 @@ impl Derive<OutputColumns> for LogicalGetBorrowed<'_> {
         Arc::new(
             self.projections()
                 .iter()
-                .map(|x| Column(self.first_column().0 + usize::from(*x)))
+                .map(|x| Column(self.first_column().0 + (*x)))
                 .collect(),
         )
     }
